+++ lib/brando/pages/page.ex
@@ -42,6 +42,7 @@ defmodule Brando.Pages.Page do
     belongs_to(:creator, Brando.User)
     belongs_to(:parent, __MODULE__)
     has_many(:children, __MODULE__, foreign_key: :parent_id)
+    has_many(:fragments, Brando.Pages.PageFragment)
     field(:meta_description, :string)
     field(:meta_keywords, :string)
     timestamps()