+++ lib/brando/pages/pages.ex
@@ -106,7 +106,7 @@ defmodule Brando.Pages do
 
     case page do
       nil -> {:error, {:page, :not_found}}
-      page -> {:ok, page}
+      page -> {:ok, Brando.repo().preload(page, :fragments)}
     end
   end
 
@@ -115,7 +115,7 @@ defmodule Brando.Pages do
 
     case page do
       nil -> {:error, {:page, :not_found}}
-      page -> {:ok, page}
+      page -> {:ok, Brando.repo().preload(page, :fragments)}
     end
   end
 
@@ -124,8 +124,10 @@ defmodule Brando.Pages do
   """
   def get_page(key, lang) when is_binary(key) do
     q =
-      from p in Page,
-        where: p.key == ^key and p.language == ^lang
+      from(p in Page,
+        where: p.key == ^key and p.language == ^lang,
+        preload: :fragments
+      )
 
     case Brando.repo().one(q) do
       nil -> {:error, {:page, :not_found}}
@@ -138,10 +140,11 @@ defmodule Brando.Pages do
   """
   def get_page_with_children(key, lang) when is_binary(key) do
     q =
-      from p in Page,
+      from(p in Page,
         join: c in assoc(p, :children),
         where: p.key == ^key and p.language == ^lang,
-        preload: :children
+        preload: [:children, :fragments]
+      )
 
     case Brando.repo().one(q) do
       nil -> {:error, {:page, :not_found}}
@@ -154,8 +157,10 @@ defmodule Brando.Pages do
   """
   def get_page(nil, key, lang) when is_binary(key) do
     q =
-      from p in Page,
-        where: p.key == ^key and p.language == ^lang
+      from(p in Page,
+        where: p.key == ^key and p.language == ^lang,
+        preload: [:fragments]
+      )
 
     case Brando.repo().one(q) do
       nil -> {:error, {:page, :not_found}}
@@ -165,9 +170,11 @@ defmodule Brando.Pages do
 
   def get_page(parent_key, key, lang) when is_binary(key) do
     q =
-      from p in Page,
+      from(p in Page,
         left_join: pp in assoc(p, :parent),
-        where: p.key == ^key and pp.key == ^parent_key and p.language == ^lang
+        where: p.key == ^key and pp.key == ^parent_key and p.language == ^lang,
+        preload: [:fragments]
+      )
 
     case Brando.repo().one(q) do
       nil -> {:error, {:page, :not_found}}
@@ -336,8 +343,9 @@ defmodule Brando.Pages do
 
     fragment =
       Brando.repo().one(
-        from p in PageFragment,
+        from(p in PageFragment,
           where: p.key == ^key and p.language == ^language
+        )
       )
 
     case fragment do
@@ -372,8 +380,9 @@ defmodule Brando.Pages do
 
     fragment =
       Brando.repo().one(
-        from p in PageFragment,
+        from(p in PageFragment,
           where: p.parent_key == ^parent and p.key == ^key and p.language == ^language
+        )
       )
 
     case fragment do